Estruturando o back end criando migrations com knex
