+++ backEnd/src/server.ts
@@ -1,27 +1,9 @@
 import express, { request, response } from "express";
+import routes from './routes'
 
 const app = express();
 
-app.get('/products', (request, response) => {
-    console.log('listagem de produtos');
-
-    return response.json([
-        'Rafael',
-        'Michelle',
-        '...',
-        'Lorena',
-        'Pietro',
-        'Lucas'
-    ])
-});
-
-app.post('/products', (request, response) => {
-    const product = {
-        name: 'macarrão',
-        price: 2.50
-    };
-
-    return response.json(product)
-});
+app.use(express.json());
+app.use(routes);
 
 app.listen(3333);
